+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.2.0]
+### Added
+- A way to set custom vertical offsets via android:dropDownVerticalOffset and by setting Gravity.BOTTOM
+
+### Changed
+- Updated Android Support Library to 27.1.1
+- Updated Kotlin to 1.2.61
+- Changed kotlin-stdlib-jre7 dependency to kotlin-stdlib-jdk7
+
 ## [2.1.0]
 ### Added
 - A way to dismiss popup menus programmatically
@@ -48,6 +57,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 Initial release of the library.
 
+[2.2.0]: https://github.com/zawadz88/MaterialPopupMenu/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/zawadz88/MaterialPopupMenu/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/zawadz88/MaterialPopupMenu/compare/1.4.2...2.0.0
 [1.4.2]: https://github.com/zawadz88/MaterialPopupMenu/compare/1.4.1...1.4.2